refactor(subtopic): move repository package to persistence.model
