+++ src/main/java/com/kllhy/roadmap/roadmap/persistence/model/ImportanceLevel.java
@@ -1,4 +1,4 @@
-package com.kllhy.roadmap.roadmap.repository;
+package com.kllhy.roadmap.roadmap.persistence.model;
 
 public enum ImportanceLevel {
     DEFAULT, RECOMMENDED, OPTIONAL, PARALLEL
