+++ src/main/java/com/kllhy/roadmap/roadmap/persistence/model/ResourceType.java
@@ -1,4 +1,4 @@
-package com.kllhy.roadmap.roadmap.repository;
+package com.kllhy.roadmap.roadmap.persistence.model;
 
 public enum ResourceType {
     VIDEO, OFFICIAL, POST, ROADMAP
